Fix a bug in group-by's combiner, and also a bug in into
group-by's use of `merge-with` caused some interesting special cases;
notably, it would fail to use the combiner identity correctly, causing
emitted results to have the wrong type. This patch replaces merge-with
with a custom implementation that starts every group's combiner phase
with a combiner identity. We also use transients for a performance
boost when the number of chunks is large.

Into would crash for large numbers of chunks, because concat causes
stack overflows. We now conj directly into the given collection.