Into: now conj's in reducer. Fix a bug where fuse didn't post-reduce.
This slightly changes the semantics of `into` because we're now doing
conj THEN into, instead of just into, but I don't think it should
matter--and the speedup from doing e.g. conj on sets in the reducer is
huge.

Also rewrote fuse to use uniform arrays throughout instead of juxt
magic, and reduce the number of functions.